Update logic of the free trial page
